Sandbox Certificate
This might not be the right place for this, but took me a bit of time to debug what was going on as the error was a bit unclear.  I was thinking about adding another error code to Java, but that seems like something you'd need to do internally - all the other docos and what not.  And this also seems like it'd only be a problem with the java sdk...